Move admin.runner.run back to only module using it.
